# docs/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Semana Santa API",
    "version": "v0",
    "description": "Semana Santa management API — procesion service includes Pasos, Route Sections, and plan finalization."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "API Gateway (development)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/hermandades/{hermandadId}/procesiones/{procesionId}/pasos": {
      "get": {
        "tags": ["paso-controller"],
        "summary": "Get ordered list of pasos for a procesion",
        "operationId": "getPasos",
        "parameters": [
          { "name": "hermandadId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "procesionId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Ordered list of pasos", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ReplacePasosResponse" } } } },
          "401": { "description": "Unauthorized — missing or invalid JWT" },
          "403": { "description": "Forbidden — cross-tenant access" },
          "404": { "description": "Procesion not found" }
        }
      },
      "put": {
        "tags": ["paso-controller"],
        "summary": "Atomically replace the ordered list of pasos",
        "operationId": "replacePasos",
        "parameters": [
          { "name": "hermandadId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "procesionId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReplacePasosRequest" } } } },
        "responses": {
          "200": { "description": "Pasos replaced", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ReplacePasosResponse" } } } },
          "400": { "description": "Invalid input (validation error or duplicate position)" },
          "401": { "description": "Unauthorized — missing or invalid JWT" },
          "403": { "description": "Forbidden — cross-tenant or titular from another hermandad" },
          "404": { "description": "Procesion not found" },
          "409": { "description": "Conflict — plan already finalized, pasos are immutable" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReplacePasosRequest": {
        "required": ["pasos"],
        "type": "object",
        "properties": {
          "pasos": { "type": "array", "items": { "$ref": "#/components/schemas/PasoItemRequest" }, "minItems": 1 }
        }
      },
      "PasoItemRequest": {
        "required": ["position", "titularId"],
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "nullable": true },
          "position": { "type": "integer", "format": "int32", "minimum": 0 },
          "titularId": { "type": "string", "format": "uuid" },
          "notes": { "type": "string", "nullable": true }
        }
      },
      "ReplacePasosResponse": {
        "type": "object",
        "properties": {
          "pasos": { "type": "array", "items": { "$ref": "#/components/schemas/PasoItemResponse" } }
        }
      },
      "PasoItemResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "procesionId": { "type": "string", "format": "uuid" },
          "position": { "type": "integer", "format": "int32" },
          "titularId": { "type": "string", "format": "uuid" },
          "notes": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Paste your Keycloak access token here.\n\nGet one via:\n  curl -s -X POST http://localhost:8180/realms/semana-santa/protocol/openid-connect/token \\\n    -d \"grant_type=password&client_id=semana-santa-client&client_secret=secret&username=qa-admin-user&password=test\" \\\n    | jq -r '.access_token'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}